requirements.txt
discord==1.7.3
Flask==2.0.1
replit==3.2.4
python-dotenv==0.19.0
ffmpeg==1.4
PyNaCl==1.4.0
youtube_dl==2021.6.6
better-profanity==0.7.0
wikipedia==1.4.0
asyncpraw==7.4.0
google-api-python-client==2.22.0
Pillow==8.3.2
googletrans==3.1.0a0
motor==2.5.1